Agendar Turno: submitAppointmentDetails v2

+++ src/main/java/com/disenio/TFI/model/AppointmentData.java
@@ -13,5 +13,5 @@ public class AppointmentData {
     private Date date;
     private String description;
     private Long patientId;
-    private Long duration;
+    private int duration;
 }
